requirements.txt
measurement
pyfunctional
beautifulsoup4
requests
plotly
cufflinks
pandas
pytest
